Don't trap on signed int overflow (-fwrapv instead of -ftrapv)
this option makes clang insert an UD2 (causes "illegal instrction"
error) when a signed integer overflows.. I'm not sure we want that.

On the one hand, int overflows for array indexes and similar are bad..
but e.g. for the seed value of a RNG they're ok.
(But then again one could probably use unsigned ints for the ok cases)
Furthermore I guess that -ftrapv causes overhead for the checks.