Start replacing CEGUI_SDLHooks with something platform-independent
The API is done, it's already used for initializing and updating +
rendering, but I'm not passing keyboard/mouse events yet.

The code is based on CEGUI_SDLHooks, but doesn't use a struct or a
keyboard map (because actually, d3bfg and CEGUI use the same key numbers
internally, isn't that awesome?).
I hope that I can avoid having to call this from platformspecific
backends and can pass all events from the generic code.