Call idCEGUI::NotifyDisplaySizeChanged() from backends, SE_MOUSE_LEAVE on SDL
Unfortunately, calling idCEGUI::NotifyDisplaySizeChanged() from
R_SetNewMode() doesn't seem to suffice, as that function is not
necessarily called when resizeing the window.

So this one hook must be called from the backends :/
It'll still be called from R_SetNewMode(), for when one changes
r_windowWidth/Height and then does a vid_restart

The SDL backend now creates SE_MOUSE_LEAVE events when the mouse leaves
the window (both SDL1.2 and SDL2). For some reason, both the SWF GUI
backend and CEGUI are interested in this.