(Hopefully) fix MSVC build by wrapping CEGUI/CEGUI.h into own wrap_cegui.h
We have our own wrap_cegui.h now that undefines some macros that conflict
with MSVC C++ standardlib headers.
It also supresses some warnings on GCC/clang.
It should always be used instead of CEGUI/CEGUI.h

It's not called cegui.h, because it's in cegui/ and for case-insensitive
filesystem (Windows, OSX) would be the same as CEGUI/CEGUI.h