Fix RPATH on Linux (and probably other Unices)
cmake set an absolute path as rpath in the OpenTechEngine binary and the
cegui shared libs.
That sucks, because the binaries won't work anywhere else then.
Now, if cegui is enabled (this is only relevant for finding the cegui
libs), the RPATH is set to $ORIGIN/lib/ so the runtime linker will
always look in the lib/ directory next to the executable.